+++ src/common/Core/Search/SearchFilter.ts
@@ -1,4 +1,4 @@
 import type { SearchResultItem } from "@common/Core";
-import type { SearchOptions } from "./SearchOptions";
+import type { SearchOptions } from "../../../renderer/Core/Search/Helpers/SearchOptions";
 
 export type SearchFilter = (options: SearchOptions) => SearchResultItem[];
