+++ src/common/Core/Search/SearchResultItemFilter.ts
