+++ src/renderer/Core/Search/Helpers/index.ts
@@ -1,2 +1,2 @@
+export * from "../../../../common/Core/Search/SearchFilter";
 export * from "./elementIsVisible";
-export * from "./SearchFilter";
